+++ packages/angular-workspace/ok-angular/fv/splitter/public-api.ts
@@ -1,0 +1,2 @@
+export * from './ok-fv-splitter.directive';
+export * from './ok-fv-splitter.module';