+++ packages/angular-workspace/ok-angular/fv/sticky-header/ok-fv-sticky-header.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { OkFvStickyHeaderDirective } from './ok-fv-sticky-header.directive';
+
+import '@ni/ok-components/dist/esm/fv/sticky-header';
+
+@NgModule({
+    declarations: [OkFvStickyHeaderDirective],
+    imports: [CommonModule],
+    exports: [OkFvStickyHeaderDirective]
+})
+export class OkFvStickyHeaderModule { }