+++ packages/react-workspace/ok-react/src/fv/sticky-header/index.ts
@@ -1,0 +1,8 @@
+'use client';
+
+import { FvStickyHeader, fvStickyHeaderTag } from '@ni/ok-components/dist/esm/fv/sticky-header';
+import { wrap } from '../../utilities/react-wrapper';
+
+export { fvStickyHeaderTag };
+export { type FvStickyHeader };
+export const OkFvStickyHeader = wrap(FvStickyHeader);